Added setting 'remove_duplicate_transactions'
This setting controls whether or not to remove duplicate transactions.
If you download transactions from 2 accounts that include transactions between
then, you should use the default 'true' as otherwise the transactions appear
twice which results to uncorrect balances. Only the transaction with the amount
>0 is used, the other one is calculated automatically.

However, when you don't download both transactions, you need to switch this
setting to 'false' in order to have a correct balance on your account.